style: Reduction of unnecessary static_cast

+++ src/natives/common.cpp
@@ -12,7 +12,7 @@ using spdlog::level::to_string_view;
  */
 static cell_t LogLevelToName(SourcePawn::IPluginContext *ctx, const cell_t *params) noexcept
 {
-    auto lvl = log4sp::num_to_lvl(static_cast<int>(params[3]));
+    auto lvl = log4sp::num_to_lvl(params[3]);
     auto name = to_string_view(lvl);
 
     size_t bytes{0};
@@ -25,7 +25,7 @@ static cell_t LogLevelToName(SourcePawn::IPluginContext *ctx, const cell_t *para
  */
 static cell_t LogLevelToShortName(SourcePawn::IPluginContext *ctx, const cell_t *params) noexcept
 {
-    auto lvl = log4sp::num_to_lvl(static_cast<int>(params[3]));
+    auto lvl = log4sp::num_to_lvl(params[3]);
     auto name = to_short_c_str(lvl);
 
     size_t bytes{0};
